Fix error checking for tango.stamp versus target file.
In particular, the tango stamp is based on the time attached to the
youngest source file, which is also then attached to that source
file's associated target file. So you cannot require that the
tango.stamp is always strictly younger than all target files.

(Another potential fix for this would be to stamp the tango.stamp
afresh on every build, rather than attaching the timestamp associated
with the youngest source file. But lets try this for now, it seems
like a more stable option for the long term.)