Update tango to work with current rustc 1.2.0-nightly (2f5683913 2015-06-18).
This is largely me working around the removal of the
`Metadata::modified` method. The approach I ended up taking (after
some missteps, artifacts of which are probably still embedded in this
commit) was to factor out a `timestamp` submodule. It currently will
only work on `cfg(unix)` targets, but hopefully `mod timestamp` can be
adapted to work elsewhere.