Removed reference to unused (deprecated) path_ext feature.
Fixed test runner code to be more robust when initially creating the
out_path; multiple instances of the runner may be running in parallel,
and they interfere with each other (all trying to create the same
output directory). To deal with this, accept a fixed number of errors,
retrying after each one.

(We have to retry, rather than assuming that the interference means
that the whole directory was built, because create_dir_all may
encounter interference (and error out) during the construction of some
parent path.)