Unfinished second entry point to emit rerun-if-changed (mentioned in issue 12).
The second entry point (`process_root_and_emit_rerun`):
1) gets the path to tango
2) writes the rerun-if-changed lines
3) calls `process_root`

It currently misuses cargo's CARGO_MANIFEST_LINKS environment variable to get the path to tango: If the crate calling tango has the path to tango saved as the CARGO_MANIFEST_LINKS environment variable in it's cargo.toml, tango can access that and write the rerun-if-changed lines with the correct path to tango's files.  Needs a better way to do this.