Write rerun-if from gather_inputs, add field to Context struct.
Rerun-if-changed=PATH is emitted from gather_inputs function, for each .rs file as it is being processed.  The context struct has a field (`emit-rerun-if`) that gather_inputs checks to determine if rerun-if-changed should be emitted.

The second entry point has been removed, a config struct will be used
instead (not added yet).